Promote dScMgRoulette_c to one translation unit (ov006, 40/40)
Folds the 40 per-symbol shards covering ov006 .text 0x0210788c..0x0210a400
into src/actors/dScMgRoulette_c.cpp. Whole promotion, no held-out subset:
tubuild verify reports 40/40 MATCH, objisolate clean, reloc destinations
clean, and all 40 functions in ROM-ascending emission order. The recorded
linkcheck is SCRATCH-LINK-VERIFIED: 106/106 modules exact, 100.000000% of
compared bytes, ROM sha256 identical to the baseline control, 9 pre-existing
symbol errors with 0 new.

The destructor moves inline, in-class, declared first. The cartridge puts D1
at 0x0210788c BELOW D0 at 0x02107920 and carries no D2, which in mwccarm
2004/b56 is reachable only from an inline in-class definition; out of line it
emits D2, D0, D1 plus a D2 the ROM lacks. That also moves the key function to
InitResources, so this TU emits _ZTV/_ZTI/_ZTS -- 13 vague-linkage symbols,
2 * (5 ancestors + self) + 1 vtable, every one of them with a configured ROM
home elsewhere, hence all deadstrip-data and externalized_output empty.

The queue's pragma:1 blocker is retired by measurement, not by carrying the
pragmas: with both shard pragmas DELETED the TU still compiles 40/40, and a
file-global optimize_for_size (23/40, 17 DIFF) followed by the same pragma
bracketed around one of those 17 size-sensitive members (40/40 -- the bracket
did not bind) shows mwccarm treats it as file-global last-wins, exactly as
notes/agents/roles/writer.md says. Repeating that bracket under
`#pragma defer_codegen off` DOES bind it (39/40, the DIFF being exactly the
bracketed member), which extends the dScMgMemory2_c writer's finding to a
third pragma -- but the same flag reverses emission order, so it is not used
here. Full write-up in the manifest's boundary_evidence.
