+++ config/arm9/overlays/ov006/delinks.txt
@@ -4410,165 +4410,9 @@ src/d_s_mg_panel.c:
     complete
     .text start:0x02107858 end:0x0210788c
 
-src/_ZN15dScMgRoulette_cD1Ev.cpp:
+src/actors/dScMgRoulette_c.cpp:
     complete
-    .text start:0x0210788c end:0x02107920
-
-src/_ZN15dScMgRoulette_cD0Ev.cpp:
-    complete
-    .text start:0x02107920 end:0x021079c8
-
-src/func_ov006_021079c8.c:
-    complete
-    .text start:0x021079c8 end:0x021079cc
-
-src/func_ov006_021079cc.c:
-    complete
-    .text start:0x021079cc end:0x02107a6c
-
-src/func_ov006_02107a6c.cpp:
-    complete
-    .text start:0x02107a6c end:0x02107b14
-
-src/func_ov006_02107b14.c:
-    complete
-    .text start:0x02107b14 end:0x02107b70
-
-src/func_ov006_02107b70.c:
-    complete
-    .text start:0x02107b70 end:0x02107b94
-
-src/func_ov006_02107b94.c:
-    complete
-    .text start:0x02107b94 end:0x02107c50
-
-src/func_ov006_02107c50.c:
-    complete
-    .text start:0x02107c50 end:0x02107d20
-
-src/func_ov006_02107d20.c:
-    complete
-    .text start:0x02107d20 end:0x02107d58
-
-src/func_ov006_02107d58.c:
-    complete
-    .text start:0x02107d58 end:0x02107d80
-
-src/func_ov006_02107d80.cpp:
-    complete
-    .text start:0x02107d80 end:0x02107db8
-
-src/func_ov006_02107db8.cpp:
-    complete
-    .text start:0x02107db8 end:0x02107ea8
-
-src/func_ov006_02107ea8.c:
-    complete
-    .text start:0x02107ea8 end:0x021082c4
-
-src/func_ov006_021082c4.c:
-    complete
-    .text start:0x021082c4 end:0x021082fc
-
-src/func_ov006_021082fc.c:
-    complete
-    .text start:0x021082fc end:0x02108508
-
-src/func_ov006_02108508.c:
-    complete
-    .text start:0x02108508 end:0x02108524
-
-src/func_ov006_02108524.c:
-    complete
-    .text start:0x02108524 end:0x0210858c
-
-src/func_ov006_0210858c.c:
-    complete
-    .text start:0x0210858c end:0x021085c0
-
-src/func_ov006_021085c0.c:
-    complete
-    .text start:0x021085c0 end:0x02108650
-
-src/func_ov006_02108650.c:
-    complete
-    .text start:0x02108650 end:0x02108b90
-
-src/func_ov006_02108b90.c:
-    complete
-    .text start:0x02108b90 end:0x02108cc0
-
-src/func_ov006_02108cc0.c:
-    complete
-    .text start:0x02108cc0 end:0x02108d28
-
-src/func_ov006_02108d28.c:
-    complete
-    .text start:0x02108d28 end:0x02108e24
-
-src/func_ov006_02108e24.c:
-    complete
-    .text start:0x02108e24 end:0x02108f2c
-
-src/func_ov006_02108f2c.c:
-    complete
-    .text start:0x02108f2c end:0x0210927c
-
-src/func_ov006_0210927c.c:
-    complete
-    .text start:0x0210927c end:0x021092a0
-
-src/func_ov006_021092a0.c:
-    complete
-    .text start:0x021092a0 end:0x021092e8
-
-src/func_ov006_021092e8.c:
-    complete
-    .text start:0x021092e8 end:0x0210935c
-
-src/func_ov006_0210935c.c:
-    complete
-    .text start:0x0210935c end:0x021094ac
-
-src/func_ov006_021094ac.c:
-    complete
-    .text start:0x021094ac end:0x02109530
-
-src/func_ov006_02109530.c:
-    complete
-    .text start:0x02109530 end:0x021095ac
-
-src/func_ov006_021095ac.c:
-    complete
-    .text start:0x021095ac end:0x021095cc
-
-src/_ZN15dScMgRoulette_c13OnYoshiTryEatEi.cpp:
-    complete
-    .text start:0x021095cc end:0x021096c8
-
-src/_ZN15dScMgRoulette_c13OnTurnIntoEggEi.cpp:
-    complete
-    .text start:0x021096c8 end:0x0210980c
-
-src/_ZN15dScMgRoulette_c16CleanupResourcesEv.cpp:
-    complete
-    .text start:0x0210980c end:0x02109834
-
-src/_ZN15dScMgRoulette_c6RenderEv.cpp:
-    complete
-    .text start:0x02109834 end:0x02109aa0
-
-src/func_ov006_02109aa0.c:
-    complete
-    .text start:0x02109aa0 end:0x02109aac
-
-src/_ZN15dScMgRoulette_c8BehaviorEv.cpp:
-    complete
-    .text start:0x02109aac end:0x0210a194
-
-src/_ZN15dScMgRoulette_c13InitResourcesEv.cpp:
-    complete
-    .text start:0x0210a194 end:0x0210a400
+    .text start:0x0210788c end:0x0210a400
 
 src/minigames/d_s_mg_roulette.cpp:
     complete
