+++ include/decl_common.h
@@ -2506,7 +2506,11 @@ extern void func_ov006_02107b14(void);
 extern void func_ov006_02107b70(char*);
 extern void func_ov006_02107d80(char*);
 extern void func_ov006_02108524(char*);
-extern void func_ov006_0210858c(void*);
+/* func_ov006_0210858c is deliberately NOT declared here. Its recorded return
+   type was void; the cartridge's own bytes at 0x0210858c end `mov r0, #1`,
+   so it returns int. Nothing outside src/actors/dScMgRoulette_c.cpp, which
+   now defines it, references the symbol, so the wrong declaration is
+   dropped rather than corrected. */
 extern void func_ov006_02108cc0(char*);
 extern void func_ov006_021092a0(int*);
 extern void func_ov006_02109aa0(char*);
