--- src/func_ov006_021079c8.c
@@ -1,3 +1,0 @@
-void func_ov006_021079c8(void)
-{
-}
