--- src/func_ov006_02107d20.c
@@ -1,11 +1,0 @@
-void func_ov006_02107d20(int *p, int val)
-{
-    int i;
-    *(short *)((char *)p + 0x16) += 1;
-    for (i = 0; i < 5; i++) {
-        if (p[i] == 0) {
-            p[i] = val;
-            return;
-        }
-    }
-}
