--- src/func_ov006_02108508.c
@@ -1,3 +1,0 @@
-struct S { int w[2]; };
-extern struct S data_ov006_0213e2e8;
-void func_ov006_02108508(char *p) { *(struct S *)(p + 0x0) = data_ov006_0213e2e8; }
