--- src/func_ov006_021095ac.c
@@ -1,8 +1,0 @@
-extern void func_ov006_02109530(int* a, int* b, int c);
-
-void func_ov006_021095ac(int* a, int* b)
-{
-    a[8] = b[0];
-    a[9] = b[1];
-    func_ov006_02109530(a, b, 0x80);
-}
